refactor: extract bending-tilt selection helpers to bt_selection.py
Moved vertex and group selection helpers (_collect_preset_rows, _collect_group_rows, _base_term_region_zero_rows, _interior_mask_leaflet) to modules/energy/bt_selection.py. Preserved exact caching and selection logic. Verified with J0 preset and base-term boundary group tests.
